+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Add conversion checks to all slots in `AnnData2SCE()` (See #45)
 * Enable return conversion of `varm` in `SCE2AnnData()` (Fixes #43)
+* Store `X_name` in `AnnData2SCE()` for use by `SCE2AnnData()` and add an
+  `X_name` argument to `AnnData2SCE()` and `readH5AD()` (Fixes #7)
 
 ## zellkonverter 1.1.7 (2021-04-30)